Phase 2: extract alert console and channel helpers into submodules
